Use bundle-specific campaign event constant
Renamed the MauticMauldinClicksEvents class to ClickEvents as it is
common practice to use short classnames for event type
definitions. The bundle namespace is sufficient for differentiating
against possible classname collisions.

The CampaignSubscriber now listens for the event named in the bundle;
and the registered CampaignBuilder step will trigger that event.