Adjust namespaces of services and translations
Since our bundle is a third-party plugin and not (yet?) an accepted
addition to the Mautic core, it is probably wise to use our own
namespace for services and translations.

The use of 'mautic.page' and 'mautic.email' vendor.bundle prefixes
have been replaced with 'mauldin.click'. The 'mautic.page' translation
string references have been copied from the PageBundle and are now
used for the label and tooltip properties of our form type.

The formtype's alias has been prefixed with 'mauldin_', and the
return of its getName() method has been updated to match.