Rename Campaign Event type w/ mauldin prefix
The "event type", of the decision that is triggered within campaigns
for email link clicks, is the only identifier used to distinguish our
event from others in the 'campaign_events' DB table. Since that type
is specific to our plugin, we should avoid the 'email' prefix used by
the core. A new constant property, CampaignSubscriber::CLICK_TYPE, now
holds the string used to identify that event ('mauldin.click_link')
and is used throughout the class where there was previously the string
'email.click_link'.

Added a message string for the label of click entries in the campaign
decision list.  The descriptive label for each item in the campaign
event list, in the 'MauticCampaignBundle:Campaign:events.html.php'
view, uses a different message than those defined for the campaign
builder. The message name is composed of a hard-coded
prefix ('mautic.campaign.')  followed by the event's 'type'
identifier--'mauldin.click_link' in this case.