Define QueuedTransportInterface for execution test
The Brick Abode core changes rely on testing if a mailer transport
object has a property named 'rabbitmqidproperty'. That property is
defined on our RabbitmqTransport for the sake of this test, but is
otherwise unused. In modern PHP, it is more common to test whether an
object implements an interface that provides the [un]desired
functionality. For this purpose, such an interface has been defined in
the QueuedTransportInterface. That interface defines a setter
injection method, 'setTransportQueue', which takes an instance of
TransportQueueInterface as its sole argument. There's also a getter
which will throw an exception if the queue property is not bound. The
queue's interface has also been defined, and implementation thereof
will be used to provide our transport with access to the message queue
functionality.