Override modified core services via compiler pass
The initial implementation of this plugin by Brick Abode relied on
changes to the core MailHelper and EmailModel classes, wherein the
mailer transport is checked for use of an AMQP message queue.

In the interest of avoiding direct modification of core classes, this
bundle introduces two new classes which extend their respective core
classes and define method overrides as appropriate. A CompilerPass has
been added to the bundle load method which replaces the instantiated
class with our extended implementation for each of the relevant
application services.

The overridden methods--as now provided by the plugin--differ from
earlier core alterations in that the mailer transport is now tested
for implementation of the QueuedTransportInterface rather than the
existence of an otherwise unused class property.