Implement QueuedTransportInterface on RabbitmqTransport
This enables the short circuit in the QueuedMailHelper and
QueuedEmailModel services provided by the plugin.

Removed constructor override in favor of setter injection for
dependency. The AMQPStreamConnection property has been replaced with a
TransportQueueInterface which handles the queue declaration.

The TransportQueue wraps the channel publish method and handles
casting the message to an AMQPMessage, so the transport's 'send()'
body has been updated to simply pass the serialized message.

The Transport's 'sendDirect()' method now returns the result of the
parent send method as indicated by the '@return' docBlock item.

Updated the transport service definition with queue.

Defined 'mauldin.transport_queue.rabbitmq' and replaced obsolete
constructor arg on 'mautic.transport.rabbitmq' with methodCall to
setTransportQueue().