Update EventModelExtended to use ChannelHelper
Prior to this update, the two campaign event handling methods,
'consumeStartingEvents' and 'triggerStartingEventsSelect' received an
AMQPChannel as an argument. That argument has been removed from the
signatures of those methods, which are now identical to the signatures
of the core methods they reproduce. The subtlety of that argument's
existence was compounded by the omission of corresponding '@param'
comments in the methods' documentation.

To provide equivalent functionality in the absence of the channel as
an argument, the EventModelExtended now has a facility for binding the
newly defined ChannelHelper via setter injection. Additional methods
have been defined within the model class to reduce repetition in
bootstrapping and using the channel and its queues.

One such repetition was the declaration of per-campaign queues. Each
call to AMQPChannel::queue_declare() received a queue name argument
which was the concatenation of a hard-coded string prefix and a
campaign ID. That prefix string now exists as a class constant and the
model class has a method for retrieving a QueueReference object when
called with the campaign ID alone.

This commit introduces another significant method signature change in
the removal of the getCampaignLeadIdsNext method's first argument, $q,
a Doctrine QueryBuilder. Since that variable is used solely within the
method to which it was passed, there doesn't appear to be any reason
to define it in the calling scope (which is complex enough as is). The
QueryBuilder is now instantiated within its narrowest scope of use.
The method is now prefaced with a doc block.

Config update: added setChannelHelper method call to definition for
the EventModelExtended service.

N.B. Minor code style changes have been introduced throughout the
EventModelExtended class by the php-cs-fixer pre-commit hook.