Update Trigger* commands
1. TriggerCampaignCommand

Renamed the $model variable to $eventModel to clarify distinction
between that and $campaignModel, the other model variable within the
method scope.

Removed $connection and $channel variables, as that dependency is now
injected into the EventModelExtended.

Removed the now defunct $channel argument passed to calls of
EventModelExtended::triggerStartingEventsSelect().

2. TriggerConsumerCampaignCommand

Removed unused '--scheduled-only' and '--negative-only' options.

Renamed $model to $eventModel as above.

Removed $connection and $channel variables as above. The $channel
argument is not passed to $eventModel::consumeStartingEvents()
as that argument no longer exists. Note that there is a new variable,
within the execute() method scope, named $channel (described below).

Renamed $max variable to $maxEvents.

Added '--max-retries' command option (default: 10), the value of which
is bound to the $maxRetries variable used in place of the previously
hard-coded $max_timeout variable.

Deleted the unused $process_list array.

Renamed snake-cased variables to lower-camel-case per Symfony Coding
Standards / Naming Conventions:
http://symfony.com/doc/current/contributing/code/standards.html#naming-conventions

Updated the channel callback waiting logic to use the instance of
QueueChannel bound to the EventModelExtended object. The object is
referenced therein by the variable $channel, but it is not an
instance of AMQPChannel as before. The QueueChannel provides a
shorthand wait() method in which the timeout value is passed as the
first argument.

Adjusted some bad indentation to match the logical structure of the
code.