Fix queue timeout typos & argument bugs
The ProcessEmailQueueCommand's timeout option variable declaration had
a typo. It was declared as '$timoutPeriod', but referenced as
'$timeoutPeriod' as intended.

Also replaced the addition operator (+) with a concatenation
operator (.) in the timeout output line writing string construction.

The RabbitmqTransportQueue and QueueReference classes were forwarding
arguments to the QueueChannel's wait() method in the order used by
AMQPChannel, instead of the order in the shorthand method defined on
that wrapper.