Feature: record AB-Test results when rolling out and improve stats
This is intended to allow seeing the AB-Test results after the winner
has been picked. The original Mautic behavior for Emails was to reset
the stats.

This depends on: https://github.com/MauldinEconomics/csi-unicorn/pull/93
which adds a new table for storing the AB-Test results.

The stats showed in the details pages are also improved for Emails and
Pages.