Bugfix: workaround for newlycreated lead flag
This is necessary for the FoF cookie integration in the
MauticMauldinCSIBundle plugin.

A lead is created with no email upon a page hit, and the subsequent form
submission only updates it with the email.

So the "newly created" flag is not set on the form submission.

Since we rely on this flag to capture the FoF tracking cookies, we make
this workaround.