Adjust queueing condition in requestCacheUpdate
The 'apiCall' method will throw an exception on unsuccessful SET API
responses, e.g. when an attempt is made to get the status of a
nonexistent build. This exception was uncaught in SETRequestModel's
requestCacheUpdate method, thereby preventing the initial queueing of
builds. Furthermore, that exception was uncaught in the SETListModel
method of the same name, also preventing upsertSetList on previously
unqueued lists.

The SETRequestModel::getTimeFinished method has been updated to return
an actual `false` for queued-but-unfinished builds; and the call to
that method in requestCacheUpdate has been wrapped in a try...catch,
whereby an exception will be considered cause for requesting the list
to be queued.

Also, the DocBlock comments on getTimeFinished have been reformatted
and the @return tag now specifies the correct return types. Note that
strtotime does not return a DateTime instance:

    int strtotime ( string $time [, int $now = time() ] )

-- http://php.net/manual/en/function.strtotime.php