Move 'Create an A/B test' div to tab-pane
With the "Contacts" pane always being shown, and having been marked-up
as a tab pane whether (showVariants || showTranslations) or not,
rendering the "Create an A/B test" div not as a 'tab-pane' created
some oddity in the markup. Since the tabs nav UL was always rendered
whether there were variants and translations or not, there might as
well be a tab for the A/B test link. This also allows for the
avoidance of duplicate markup for the new "Contacts" tab-pane, and for
that pane to have a corresponding tab in either rendering scenario.

Added a translation key, 'mautic.core.ab_test' for the tab label.

Also fixed the use of a short preprocessor open tag, in the conditional
statement that controls tab rendering, which was not necessarily safe
in stock PHP environments--presumably why the rest of the templates
use the full opening syntax.