Define ABTestResult entity & repository
The VariantModelTrait saves the results of AB tests to the database,
but is doing so directly using DBAL. This commit defines an entity
to similarly represent that data, differing in the table name and
addition of a dedicated ID field.

Custom EntityRepository class defines override for magic
'findOneByEntity' call, as "entity" is not a propery of
ABTestResult (mapped or otherwise), though it is used as a shortcut
for setting the mapped 'entityId' and 'entityType' values.