Expand blank message rejection condition
Although it is VERY likely the payload will be a string, the
AMQPMessage class does not explicitly cast its setBody() argument to a
string, so it is maybe possible that something else (like NULL) could
make it into the body property. So instead of testing that the
"payload" is identical to the empty string, the condition now just
tests whether it is any kind of empty.

Also fixed the formatting of the conditional, which was not PSR-2
compliant as it was missing a space between the keyword and the
opening parenthesis.