Dispatch MailSendResponseEvent after SendGridApiFacade sends message
The purpose of this event is to allow plugins to record information
contained within SendGrid API send responses.

Defined a new SendGrid Event class, MailSendResponseEvent. The
dispatched event instance holds a SendGrid\Response object in addition
to the 'mail' and 'message' properties of its parent class, the
GetMailMessageEvent.

The event is dispatched by the SendGridApiFacade after a SendGrid Mail
has been sent and the response has been verified by the
SendGridApiResponse service.